DATAJPA-1235 - Polishing.
Original pull request: #244.
