+++ src/main/java/org/springframework/data/jpa/repository/query/EmptyDeclaredQuery.java
@@ -15,16 +15,17 @@
  */
 package org.springframework.data.jpa.repository.query;
 
-import org.springframework.lang.Nullable;
-import org.springframework.util.Assert;
-
 import java.util.Collections;
 import java.util.List;
 
+import org.springframework.lang.Nullable;
+import org.springframework.util.Assert;
+
 /**
- * NULL-Object pattern implementation.
+ * NULL-Object pattern implementation for {@link DeclaredQuery}.
  *
  * @author Jens Schauder
+ * @since 2.0.3
  */
 class EmptyDeclaredQuery implements DeclaredQuery {
 
@@ -33,36 +34,64 @@ class EmptyDeclaredQuery implements DeclaredQuery {
 	 */
 	static final DeclaredQuery EMPTY_QUERY = new EmptyDeclaredQuery();
 
+	/*
+	 * (non-Javadoc)
+	 * @see org.springframework.data.jpa.repository.query.DeclaredQuery#hasNamedParameter()
+	 */
 	@Override
 	public boolean hasNamedParameter() {
 		return false;
 	}
 
+	/*
+	 * (non-Javadoc)
+	 * @see org.springframework.data.jpa.repository.query.DeclaredQuery#getQueryString()
+	 */
 	@Override
 	public String getQueryString() {
 		return "";
 	}
 
+	/*
+	 * (non-Javadoc)
+	 * @see org.springframework.data.jpa.repository.query.DeclaredQuery#getAlias()
+	 */
 	@Override
 	public String getAlias() {
 		return null;
 	}
 
+	/*
+	 * (non-Javadoc)
+	 * @see org.springframework.data.jpa.repository.query.DeclaredQuery#hasConstructorExpression()
+	 */
 	@Override
 	public boolean hasConstructorExpression() {
 		return false;
 	}
 
+	/*
+	 * (non-Javadoc)
+	 * @see org.springframework.data.jpa.repository.query.DeclaredQuery#isDefaultProjection()
+	 */
 	@Override
 	public boolean isDefaultProjection() {
 		return false;
 	}
 
+	/*
+	 * (non-Javadoc)
+	 * @see org.springframework.data.jpa.repository.query.DeclaredQuery#getParameterBindings()
+	 */
 	@Override
 	public List<StringQuery.ParameterBinding> getParameterBindings() {
 		return Collections.emptyList();
 	}
 
+	/*
+	 * (non-Javadoc)
+	 * @see org.springframework.data.jpa.repository.query.DeclaredQuery#deriveCountQuery(java.lang.String, java.lang.String)
+	 */
 	@Override
 	public DeclaredQuery deriveCountQuery(@Nullable String countQuery, @Nullable String countQueryProjection) {
