+++ src/main/java/org/springframework/data/jpa/repository/query/package-info.java
@@ -1,5 +1,5 @@
 /**
- * Query implementation to exectue queries against JPA.
+ * Query implementation to execute queries against JPA.
  */
 @org.springframework.lang.NonNullApi
 package org.springframework.data.jpa.repository.query;
